requirements.txt
absl-py==0.9.0
apturl==0.5.2
asn1crypto==0.24.0
astor==0.8.1
astroid==2.3.3
attrs==19.3.0
autopep8==1.4.4
backcall==0.1.0
bleach==3.1.0
Brlapi==0.6.6
celluloid==0.2.0
certifi==2018.1.18
chardet==3.0.4
Click==7.0
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cvxopt==1.2.3
cycler==0.10.0
dash==1.9.0
dash-core-components==1.8.0
dash-html-components==1.0.2
dash-renderer==1.2.4
dash-table==4.6.0
decorator==4.4.1
defer==1.0.6
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
entrypoints==0.3
Flask==1.1.1
Flask-Compress==1.4.0
future==0.18.2
gast==0.3.3
get==2019.4.13
google-pasta==0.1.8
grpcio==1.27.2
gunicorn==20.0.4
h5py==2.10.0
httplib2==0.9.2
idna==2.6
imageio==2.6.1
importlib-metadata==1.3.0
imutils==0.5.3
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.2
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.0.2
nbconvert==5.6.1
nbformat==5.0.3
netifaces==0.10.4
networkx==2.4
notebook==6.0.2
numpy==1.18.1
oauth==1.0.1
olefile==0.45.1
opencv-contrib-python==4.1.2.30
opencv-python==4.1.2.30
pandocfilters==1.4.2
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.0.0
plotly==4.5.0
post==2019.4.13
prometheus-client==0.7.1
prompt-toolkit==3.0.2
protobuf==3.11.3
ptyprocess==0.6.0
public==2019.4.13
pycairo==1.16.2
pycodestyle==2.5.0
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.5.2
pygobject==3.26.1
pylint==2.4.4
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.6
pyRFC3339==1.0
pyrsistent==0.15.6
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2018.3
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==3.12
pyzmq==18.1.1
query-string==2019.4.13
reportlab==3.4.0
request==2019.4.13
requests==2.18.4
requests-unixsocket==0.1.5
retrying==1.3.3
rope==0.14.0
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.13.2
six==1.14.0
sklearn==0.0
system-service==0.3
systemd-python==234
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
wadllib==1.3.2
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
wrapt==1.12.0
xkit==0.0.0
zipp==0.6.0
zope.interface==4.3.2
